# guide/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Guide Reporting App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LoginRequestRole = typeof LoginRequestRole[keyof typeof LoginRequestRole];


export const LoginRequestRole = {
  guide: 'guide',
  admin: 'admin',
} as const;

export interface LoginRequest {
  phone: string;
  role: LoginRequestRole;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  guide: 'guide',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  phone: string;
  name: string;
  role: UserRole;
  dailyRate: number;
  createdAt: string;
}

export type TripStatus = typeof TripStatus[keyof typeof TripStatus];


export const TripStatus = {
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Trip {
  id: number;
  name: string;
  location: string;
  startDate: string;
  endDate: string;
  leadGuideId: number;
  status: TripStatus;
  createdAt: string;
}

export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];


export const AttendanceStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  location_mismatch: 'location_mismatch',
  incomplete: 'incomplete',
} as const;

export interface Attendance {
  id: number;
  guideId: number;
  tripId: number;
  date: string;
  checkInTime?: string | null;
  checkInLatitude?: number | null;
  checkInLongitude?: number | null;
  checkInLocationName?: string | null;
  checkInSelfieUrl?: string | null;
  checkInDistance?: number | null;
  checkOutTime?: string | null;
  checkOutLatitude?: number | null;
  checkOutLongitude?: number | null;
  checkOutLocationName?: string | null;
  checkOutSelfieUrl?: string | null;
  checkOutDistance?: number | null;
  notes?: string | null;
  status: AttendanceStatus;
  verifiedAt?: string | null;
  verifiedBy?: number | null;
  createdAt: string;
}

export interface AdminDashboard {
  activeTrips: number;
  totalGuides: number;
  todayCheckIns: number;
  missingCheckIns: number;
  locationMismatchFlags: number;
}

export type AdminGuideTodayStatus = typeof AdminGuideTodayStatus[keyof typeof AdminGuideTodayStatus];


export const AdminGuideTodayStatus = {
  checked_in: 'checked_in',
  checked_out: 'checked_out',
  missing: 'missing',
  idle: 'idle',
} as const;

export interface AdminGuide {
  id: number;
  name: string;
  phone: string;
  activeTripName: string | null;
  todayStatus: AdminGuideTodayStatus;
  lastCheckInTime: string | null;
  flagged: boolean;
  daysLogged: number;
}

export type AdminAttendanceLogStatus = typeof AdminAttendanceLogStatus[keyof typeof AdminAttendanceLogStatus];


export const AdminAttendanceLogStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  location_mismatch: 'location_mismatch',
  incomplete: 'incomplete',
} as const;

export interface AdminAttendanceLog {
  id: number;
  guideName: string;
  tripName: string;
  date: string;
  checkInTime?: string | null;
  checkInLatitude?: number | null;
  checkInLongitude?: number | null;
  checkInLocationName?: string | null;
  checkInSelfieUrl?: string | null;
  checkInDistance?: number | null;
  checkOutTime?: string | null;
  checkOutLatitude?: number | null;
  checkOutLongitude?: number | null;
  checkOutLocationName?: string | null;
  checkOutSelfieUrl?: string | null;
  checkOutDistance?: number | null;
  notes?: string | null;
  status: AdminAttendanceLogStatus;
}

export type VerifyAttendanceRequestStatus = typeof VerifyAttendanceRequestStatus[keyof typeof VerifyAttendanceRequestStatus];


export const VerifyAttendanceRequestStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface VerifyAttendanceRequest {
  attendanceId: number;
  status: VerifyAttendanceRequestStatus;
}

export interface AdminReport {
  guideId: number;
  guideName: string;
  tripName: string | null;
  totalTripDays: number;
  approvedDays: number;
  rejectedDays: number;
  pendingDays: number;
  incompleteDays: number;
  locationMismatchDays: number;
  payableAmount: number;
}

export type AdminTripStatus = typeof AdminTripStatus[keyof typeof AdminTripStatus];


export const AdminTripStatus = {
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface AdminTrip {
  id: number;
  name: string;
  location: string;
  startDate: string;
  endDate: string;
  leadGuideId: number;
  leadGuideName: string;
  status: AdminTripStatus;
}

export type CheckInBody = {
  latitude: number;
  longitude: number;
  locationName: string;
  selfie: Blob;
};

export type CheckOutBody = {
  latitude: number;
  longitude: number;
  locationName: string;
  selfie: Blob;
};