# guide/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Guide Reporting App
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      description: Authenticates a guide or admin by phone number and role
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /trips/active:
    get:
      operationId: getActiveTrip
      tags: [trips]
      summary: Get active trip
      description: Returns the currently active trip for the logged-in guide
      responses:
        "200":
          description: Active trip found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "204":
          description: No active trip currently assigned
  /attendance/today:
    get:
      operationId: getTodayAttendance
      tags: [attendance]
      summary: Get today's attendance
      description: Returns today's check-in/check-out log for the logged-in guide
      responses:
        "200":
          description: Today's attendance details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "204":
          description: Not checked in yet today
  /attendance/check-in:
    post:
      operationId: checkIn
      tags: [attendance]
      summary: Check in
      description: Submits today's check-in GPS location and selfie
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
                locationName:
                  type: string
                selfie:
                  type: string
                  format: binary
              required: [latitude, longitude, locationName, selfie]
      responses:
        "200":
          description: Checked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: Already checked in or bad request
  /attendance/check-out:
    post:
      operationId: checkOut
      tags: [attendance]
      summary: Check out
      description: Submits today's check-out GPS location and selfie
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                latitude:
                  type: number
                longitude:
                  type: number
                locationName:
                  type: string
                selfie:
                  type: string
                  format: binary
              required: [latitude, longitude, locationName, selfie]
      responses:
        "200":
          description: Checked out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: Not checked in today, already checked out, or bad request
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    LoginRequest:
      type: object
      properties:
        phone:
          type: string
        role:
          type: string
          enum: [guide, admin]
      required:
        - phone
        - role
    User:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [guide, admin]
        dailyRate:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - phone
        - name
        - role
        - dailyRate
        - createdAt
    Trip:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        leadGuideId:
          type: integer
        status:
          type: string
          enum: [active, completed, cancelled]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - location
        - startDate
        - endDate
        - leadGuideId
        - status
        - createdAt
    Attendance:
      type: object
      properties:
        id:
          type: integer
        guideId:
          type: integer
        tripId:
          type: integer
        date:
          type: string
        checkInTime:
          type: string
          format: date-time
          nullable: true
        checkInLatitude:
          type: number
          nullable: true
        checkInLongitude:
          type: number
          nullable: true
        checkInLocationName:
          type: string
          nullable: true
        checkInSelfieUrl:
          type: string
          nullable: true
        checkInDistance:
          type: integer
          nullable: true
        checkOutTime:
          type: string
          format: date-time
          nullable: true
        checkOutLatitude:
          type: number
          nullable: true
        checkOutLongitude:
          type: number
          nullable: true
        checkOutLocationName:
          type: string
          nullable: true
        checkOutSelfieUrl:
          type: string
          nullable: true
        checkOutDistance:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, rejected, location_mismatch, incomplete]
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        verifiedBy:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - guideId
        - tripId
        - date
        - status
        - createdAt
...